N20 Motors from Adafruit
These motors have a hall-effect sensor that can tell which direction and how fast a motor is turning.  N20Motor.py is a module that has Motor and Wheel classes just like DriveMotor.py but they are specific to this motor and have additional features.

Not tested yet!

In addition, the specs for the motors are included for reference.